requirements.txt
ccxt
pandas==2.0.2
ta-lib==0.4.24
torch==2.0.1
scikit-learn==1.2.2
backtrader==1.9.74.123
pyyaml==6.0
numpy==1.24.3
